refactor: combine translate & analyze visitors by node type
