--- packages/translator/src/declaration.ts
@@ -1,11 +1,0 @@
-import type { types as t } from "@marko/compiler";
-import * as writer from "./util/writer";
-import { isOutputHTML } from "./util/marko-config";
-
-export default function (declaration: t.NodePath<t.MarkoDeclaration>) {
-  if (isOutputHTML(declaration)) {
-    writer.writeTo(declaration)`<?${declaration.node.value}?>`;
-  }
-
-  declaration.remove();
-}
