+++ packages/translator/src/util/evaluate.ts
