+++ packages/translator/src/util/runtime.ts
@@ -1,7 +1,7 @@
 import { types as t } from "@marko/compiler";
 import { importNamed } from "@marko/babel-utils";
 import { getMarkoOpts } from "./marko-config";
-import { Reserve, getParentSectionId } from "../analyze/util/sections";
+import { Reserve, getParentSectionId } from "./sections";
 
 export function importRuntime<T extends t.Node>(
   path: t.NodePath<T>,
@@ -62,20 +62,49 @@ export function callRead(path: t.NodePath, reference: Reserve) {
     case 1:
       return callRuntime(path, "readInOwner", t.numericLiteral(reference.id));
     default:
-      return callRuntime(path, "readInOwner", t.numericLiteral(reference.id), t.numericLiteral(diff));
+      return callRuntime(
+        path,
+        "readInOwner",
+        t.numericLiteral(reference.id),
+        t.numericLiteral(diff)
+      );
   }
 }
 
-export function callQueue(path: t.NodePath, fnIdentifier: t.Identifier, reference: Reserve, value: t.Expression) {
+export function callQueue(
+  path: t.NodePath,
+  fnIdentifier: t.Identifier,
+  reference: Reserve,
+  value: t.Expression
+) {
   const sectionId = getParentSectionId(path);
   const diff = getScopeDepthDifference(reference, sectionId);
   switch (diff) {
     case 0:
-      return callRuntime(path, "queue", fnIdentifier, t.numericLiteral(reference.id), value);
+      return callRuntime(
+        path,
+        "queue",
+        fnIdentifier,
+        t.numericLiteral(reference.id),
+        value
+      );
     case 1:
-      return callRuntime(path, "queueInOwner", fnIdentifier, t.numericLiteral(reference.id), value);
+      return callRuntime(
+        path,
+        "queueInOwner",
+        fnIdentifier,
+        t.numericLiteral(reference.id),
+        value
+      );
     default:
-      return callRuntime(path, "queueInOwner", fnIdentifier, t.numericLiteral(reference.id), value, t.numericLiteral(diff));
+      return callRuntime(
+        path,
+        "queueInOwner",
+        fnIdentifier,
+        t.numericLiteral(reference.id),
+        value,
+        t.numericLiteral(diff)
+      );
   }
 }
 
@@ -85,4 +114,3 @@ function getScopeDepthDifference(reference: Reserve, sectionId: number) {
   }
   return 0;
 }
-
