+++ packages/translator/src/util/sections.ts
