+++ packages/translator/src/util/tag-name-type.ts
