+++ packages/translator/src/visitors/declaration.ts
@@ -1,0 +1,13 @@
+import type { types as t } from "@marko/compiler";
+import * as writer from "../util/writer";
+import { isOutputHTML } from "../util/marko-config";
+
+export default {
+  translate(declaration: t.NodePath<t.MarkoDeclaration>) {
+    if (isOutputHTML(declaration)) {
+      writer.writeTo(declaration)`<?${declaration.node.value}?>`;
+    }
+
+    declaration.remove();
+  },
+};
